+++ buyer/views.py
@@ -11,7 +11,9 @@ class BuyerCreateAPIView(CreateAPIView):
     permission_classes = []
 
 
-class BuyerCSVDownloadAPIView(CSVDumpAPIView):
-    bucket = settings.CSV_DUMP_BUCKET_NAME
-    key = settings.BUYERS_CSV_FILE_NAME
-    filename = settings.BUYERS_CSV_FILE_NAME
+if settings.STORAGE_CLASS_NAME == 'default':
+    # this view only works if s3 is in use (s3 is default. in local dev local storage is used)
+    class BuyerCSVDownloadAPIView(CSVDumpAPIView):
+        bucket = settings.AWS_STORAGE_BUCKET_NAME_DATA_SCIENCE
+        key = settings.BUYERS_CSV_FILE_NAME
+        filename = settings.BUYERS_CSV_FILE_NAME