+++ conf/settings.py
@@ -362,20 +362,33 @@ S3_USE_SIGV4 = env.bool('S3_USE_SIGV4', True)
 AWS_S3_HOST = env.str('AWS_S3_HOST', 's3.eu-west-1.amazonaws.com')
 
 if 'aws-s3-bucket' in VCAP_SERVICES:
-    credentials = VCAP_SERVICES['aws-s3-bucket'][0]['credentials']
+    AWS_S3_DEFAULT_BINDING_BUCKET_NAME = env.str('AWS_S3_DEFAULT_BINDING_BUCKET_NAME')
+    AWS_S3_DATA_SCIENCE_BINDING_BUCKET_NAME = env.str('AWS_S3_DATASCIENCE_BINDING_BUCKET_NAME')
+
+    bucket_credentials_map = {
+        item['name']: item
+        for item in VCAP_SERVICES['aws-s3-bucket']
+    }
+
+    default_credentials = bucket_credentials_map[AWS_S3_DEFAULT_BINDING_BUCKET_NAME]
+    datascience_credentials = bucket_credentials_map[AWS_S3_DATA_SCIENCE_BINDING_BUCKET_NAME]
+
+    # Setting up the the main S3 PaSS bucket
+    credentials = default_credentials['credentials']
     AWS_ACCESS_KEY_ID = credentials['aws_access_key_id']
     AWS_SECRET_ACCESS_KEY = credentials['aws_secret_access_key']
     AWS_STORAGE_BUCKET_NAME = credentials['bucket_name']
     AWS_S3_REGION_NAME = credentials['aws_region']
     AWS_S3_ENCRYPTION = True
     AWS_DEFAULT_ACL = None
-else:
-    AWS_S3_ENCRYPTION = False
-    AWS_ACCESS_KEY_ID = env.str('AWS_ACCESS_KEY_ID', '')
-    AWS_SECRET_ACCESS_KEY = env.str('AWS_SECRET_ACCESS_KEY', '')
-    AWS_STORAGE_BUCKET_NAME = env.str('AWS_STORAGE_BUCKET_NAME', '')
-    AWS_S3_REGION_NAME = env.str('AWS_S3_REGION_NAME', '')
-    AWS_DEFAULT_ACL = 'public-read'
+
+    # Setting up the the datascience s3 PaSS bucket
+    credentials = datascience_credentials['credentials']
+    AWS_ACCESS_KEY_ID_DATA_SCIENCE = credentials['aws_access_key_id']
+    AWS_SECRET_ACCESS_KEY_DATA_SCIENCE = credentials['aws_secret_access_key']
+    AWS_STORAGE_BUCKET_NAME_DATA_SCIENCE = credentials['bucket_name']
+    AWS_S3_REGION_NAME_DATA_SCIENCE = credentials['aws_region']
+
 # Admin proxy
 USE_X_FORWARDED_HOST = True
 SECURE_PROXY_SSL_HEADER = ('HTTP_X_FORWARDED_PROTO', 'https')
@@ -397,9 +410,8 @@ STANNP_VERIFICATION_LETTER_TEMPLATE_ID = env.str(
 DIRECTORY_FORMS_API_BASE_URL = env.str('DIRECTORY_FORMS_API_BASE_URL')
 DIRECTORY_FORMS_API_API_KEY = env.str('DIRECTORY_FORMS_API_API_KEY')
 DIRECTORY_FORMS_API_SENDER_ID = env.str('DIRECTORY_FORMS_API_SENDER_ID')
-DIRECTORY_FORMS_API_DEFAULT_TIMEOUT = env.int(
-    'DIRECTORY_API_FORMS_DEFAULT_TIMEOUT', 5
-)
+DIRECTORY_FORMS_API_DEFAULT_TIMEOUT = env.int('DIRECTORY_API_FORMS_DEFAULT_TIMEOUT', 5)
+DIRECTORY_FORMS_API_ZENDESK_SEVICE_NAME = env.str('DIRECTORY_FORMS_API_ZENDESK_SEVICE_NAME', 'api')
 
 # Verification letters sent with govnotify
 GOVNOTIFY_VERIFICATION_LETTER_TEMPLATE_ID = env.str(
@@ -601,13 +613,11 @@ DIRECTORY_HEALTHCHECK_TOKEN = env.str('HEALTH_CHECK_TOKEN')
 DIRECTORY_HEALTHCHECK_BACKENDS = [
     directory_healthcheck.backends.SingleSignOnBackend,
     healthcheck.backends.ElasticSearchCheckBackend,
-    healthcheck.backends.StannpBackend,
     # health_check.db.backends.DatabaseBackend and
     # health_check.cache.CacheBackend are also registered in
     # INSTALLED_APPS's health_check.db and health_check.cache
 ]
 
-CSV_DUMP_BUCKET_NAME = env.str('CSV_DUMP_BUCKET_NAME')
 CSV_DUMP_AUTH_TOKEN = env.str('CSV_DUMP_AUTH_TOKEN')
 BUYERS_CSV_FILE_NAME = 'find-a-buyer-buyers.csv'
 SUPPLIERS_CSV_FILE_NAME = 'find-a-buyer-suppliers.csv'