+++ core/helpers.py
@@ -16,23 +16,23 @@ from django.conf import (
 def upload_file_object_to_s3(file_object, bucket, key):
     s3 = boto3.client(
         's3',
-        aws_access_key_id=settings.AWS_ACCESS_KEY_ID,
-        aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY,
-        region_name=settings.AWS_S3_REGION_NAME,
+        aws_access_key_id=settings.AWS_ACCESS_KEY_ID_DATA_SCIENCE,
+        aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY_DATA_SCIENCE,
+        region_name=settings.AWS_S3_REGION_NAME_DATA_SCIENCE,
     )
     s3.put_object(
         Bucket=bucket,
         Key=key,
-        Body=file_object.getvalue()
+        Body=file_object.getvalue(),
     )
 
 
 def get_file_from_s3(bucket, key):
     s3 = boto3.client(
         's3',
-        aws_access_key_id=settings.AWS_ACCESS_KEY_ID,
-        aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY,
-        region_name=settings.AWS_S3_REGION_NAME,
+        aws_access_key_id=settings.AWS_ACCESS_KEY_ID_DATA_SCIENCE,
+        aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY_DATA_SCIENCE,
+        region_name=settings.AWS_S3_REGION_NAME_DATA_SCIENCE,
     )
     file_object = s3.get_object(
         Bucket=bucket,