+++ supplier/helpers.py
@@ -1,9 +1,16 @@
 import csv
+
+from directory_constants import user_roles
+from rest_framework.serializers import ValidationError
+
 from django.db.models import BooleanField, Case, Count, When, Value
 from django.utils.functional import cached_property
 
 from company.models import Company
 from supplier.models import Supplier
+
+
+MESSAGE_ADMIN_NEEDED = 'A business profile must have at least one admin'
 
 
 class SSOUser:
@@ -40,7 +47,9 @@ def generate_suppliers_csv(file_object, queryset):
         'company__ownershipinvite',
         'ownershipinvite',
         'company__collaboratorinvite',
-        'collaboratorinvite'
+        'collaboratorinvite',
+        'collaborationinvite',
+        'company__collaborationinvite',
     )
     fieldnames = [field.name for field in Supplier._meta.get_fields()
                   if field.name not in csv_excluded_fields]
@@ -80,3 +89,10 @@ def generate_suppliers_csv(file_object, queryset):
             supplier['company__sectors'] = ''
 
         writer.writerow(supplier)
+
+
+def validate_other_admins_connected_to_company(company, sso_ids):
+    # a company must have at least ope admin attached to it
+    suppliers = company.suppliers.all()
+    if suppliers.filter(role=user_roles.ADMIN).exclude(sso_id__in=sso_ids).count() == 0:
+        raise ValidationError(MESSAGE_ADMIN_NEEDED)